# api.yml
openapi: 3.0.0
info:
  title: IMDTravel Microservices (Full System)
  description: "Documentação OpenAPI completa para todos os microsserviços do sistema IMDTravel, incluindo o 'imdtravel' (façade) e os serviços 'airlineshub', 'exchange', e 'fidelity' para testes diretos."
  version: 1.0.0

# Define todos os servidores/portas. Você pode trocar entre eles no Postman.
servers:
  - url: http://localhost:8080
    description: IMDTravel
  - url: http://localhost:8081
    description: AirlinesHub
  - url: http://localhost:8082
    description: Exchange
  - url: http://localhost:8083
    description: Fidelity

# Tags para agrupar os endpoints por serviço
...
paths:
...
  /bonus:
    post:
      summary: (Fidelity) Registrar bônus para usuário
      tags: [Fidelity]
      description: Adiciona pontos de bônus a um usuário.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BonusRequest'
      responses:
        '200':
          description: Bônus registrado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BonusResponse'
        '400':
          description: 'Requisição inválida (ex: bônus <= 0).'
  /points:
    get:
      summary: (Fidelity) Consultar pontos de um usuário
      tags: [Fidelity]
      description: Retorna o total de pontos e o histórico de um usuário.
      parameters:
        - in: query
          name: user
          required: true
          schema:
            type: string
          example: "usuario-teste-123"
      responses:
        '200':
          description: Pontos do usuário (retorna 0 se usuário não existir).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPoints'
components:
  schemas:
...
    BonusRequest:
      type: object
      properties:
        user: { type: string, example: "usuario-teste-123" }
        bonus: { type: integer, example: 500 }
    BonusResponse:
      type: object
      description: Resposta do endpoint /bonus.
      properties:
        success: { type: boolean, example: true }
        user: { type: string, example: "usuario-teste-123" }
        bonus_added: { type: integer, example: 500 }
        total_points: { type: integer, example: 1500 }
    BonusRecord:
      type: object
      properties:
        User: { type: string }
        Bonus: { type: integer }
        Timestamp: { type: string, format: date-time }
    UserPoints:
      type: object
      description: Resposta do endpoint /points.
      properties:
        User: { type: string, example: "usuario-teste-123" }
        TotalPoints: { type: integer, example: 1500 }
        Records:
          type: array
          items:
            $ref: '#/components/schemas/BonusRecord'